requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
Faker==37.5.3
fonttools==4.59.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.5
myplotlib==1.6.0
narwhals==2.2.0
numba==0.61.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
streamlit==1.48.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
